bootloader/tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "BaBa0525", email = "103883300+BaBa0525@users.noreply.github.com"},
]
dependencies = [
    "pyserial>=3.5",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
